+++ src/Bitpay/Client/Response.php
@@ -57,7 +57,8 @@ class Response implements ResponseInterface
     {
         $response = new self($rawResponse);
         $lines    = preg_split('/(\\r?\\n)/', $rawResponse);
-        for ($i = 0; $i < count($lines); $i++) {
+        $linesLen = count($lines);
+        for ($i = 0; $i < $linesLen; $i++) {
             if (0 == $i) {
                 preg_match('/^HTTP\/(\d\.\d)\s(\d+)\s(.+)/', $lines[$i], $statusLine);
                 $response->setStatusCode($statusCode = $statusLine[2]);