+++ src/Bitpay/InvoiceInterface.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @license Copyright 2011-2014 BitPay Inc., MIT License 
+ * @license Copyright 2011-2014 BitPay Inc., MIT License
  * see https://github.com/bitpay/php-bitpay-client/blob/master/LICENSE
  */
 
@@ -352,4 +352,8 @@ interface InvoiceInterface
     /**
      */
     public function getRate();
+
+    /**
+     */
+    public function getToken();
 }