+++ src/Bitpay/Item.php
@@ -88,6 +88,8 @@ class Item implements ItemInterface
 
     /**
      * @inheritdoc
+     *
+     * @return float
      */
     public function getPrice()
     {
@@ -95,17 +97,17 @@ class Item implements ItemInterface
     }
 
     /**
-     * @param float $price
+     * @param mixed $price A float, integer, or en_US formatted numeric string
      *
      * @return ItemInterface
      */
     public function setPrice($price)
     {
-        if(is_string($price)){
+        if (is_string($price)) {
             $this->checkPriceFormat($price);
         }
-        $price = (float)$price;
-        $this->price = $price;
+
+        $this->price = (float)$price;
 
         return $this;
     }
@@ -145,14 +147,21 @@ class Item implements ItemInterface
      */
     public function setPhysical($physical)
     {
-        $this->physical = (boolean) $physical;
+        $this->physical = (boolean)$physical;
 
         return $this;
     }
 
+    /**
+     * Checks the new price to include BTC
+     * values with more than 6 decimals.
+     *
+     * @param string $price The price value to check
+     * @throws \Exception
+     */
     protected function checkPriceFormat($price)
     {
-        if (1 !== preg_match('/^[0-9]+?[\.,][0-9]{1,6}?$/', $price)) {
+        if (preg_match('/^[0-9]+?[\.,][0-9]{1,6}?$/', $price) !== 1) {
             throw new \Bitpay\Client\ArgumentException("Price must be formatted as a float");
         }
     }