+++ tests/Bitpay/Client/ClientTest.php
@@ -171,6 +171,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('0.0000', $invoice->getBtcPaid());
         $this->assertEquals(315.7, $invoice->getRate());
         $this->assertEquals(false, $invoice->getExceptionStatus());
+        $this->assertEquals('abcdefghijklmno', $invoice->getToken()->getToken());
     }
 
     /**
@@ -456,7 +457,24 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $adapter = $this->getMockAdapter();
         $adapter->method('sendRequest')->willReturn($response);
         $this->client->setAdapter($adapter);
+        $token = new \Bitpay\Token();
+        $token->setToken('asdfsds');
+
+        // No token/public facade
         $invoice = $this->client->getInvoice('5NxFkXcJbCSivtQRJa4kHP');
+        $this->assertSame('invoices/5NxFkXcJbCSivtQRJa4kHP', $this->client->getRequest()->getPath());
+        $this->assertInstanceOf('Bitpay\InvoiceInterface', $invoice);
+
+        // pos token/public facade
+        $this->client->setToken($token->setFacade('pos'));
+        $invoice = $this->client->getInvoice('5NxFkXcJbCSivtQRJa4kHP');
+        $this->assertSame('invoices/5NxFkXcJbCSivtQRJa4kHP', $this->client->getRequest()->getPath());
+        $this->assertInstanceOf('Bitpay\InvoiceInterface', $invoice);
+
+        // merchant token/merchant facade
+        $this->client->setToken($token->setFacade('merchant'));
+        $invoice = $this->client->getInvoice('5NxFkXcJbCSivtQRJa4kHP');
+        $this->assertSame('invoices/5NxFkXcJbCSivtQRJa4kHP?token=asdfsds', $this->client->getRequest()->getPath());
         $this->assertInstanceOf('Bitpay\InvoiceInterface', $invoice);
     }
 
@@ -572,9 +590,9 @@ class ClientTest extends \PHPUnit_Framework_TestCase
                     'getExpirationTime', 'getCurrentTime', 'getOrderId', 'getItemDesc', 'getItemCode',
                     'isPhysical', 'getBuyerName', 'getBuyerAddress1', 'getBuyerAddress2', 'getBuyerCity',
                     'getBuyerState', 'getBuyerZip', 'getBuyerCountry', 'getBuyerEmail', 'getBuyerPhone',
-                    'getExceptionStatus', 'getBtcPaid', 'getRate', 'setId', 'setUrl',
+                    'getExceptionStatus', 'getBtcPaid', 'getRate', 'getToken', 'setId', 'setUrl',
                     'setStatus', 'setBtcPrice', 'setPrice', 'setInvoiceTime', 'setExpirationTime',
-                    'setCurrentTime', 'setBtcPaid', 'setRate', 'setExceptionStatus',
+                    'setCurrentTime', 'setBtcPaid', 'setRate', 'setToken', 'setExceptionStatus',
                 )
             )
             ->getMock();