+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Changed
+- Refactored function calls out of loops
+
+## [2.2.5] - 2015-06-23
+### Fixed
+- Item::setPrice accepts en_US formatted numeric strings
+- getInvoice now works with merchant and public facades
+
+## [2.2.4] - 2015-06-04
 ### Added
 - Token functionality for client-sided pairing
 - Set payout bitcoin amount, as returned by BitPay
@@ -12,6 +21,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Documentation from RST to Markdown
 - Autoloader now loads relative to library root directory
+- Client now sets invoice tokens upon creation and retrieval
+
+## [2.2.3] - 2015-05-29
+### Fixed
+- BitPay API errors are now passthrough
 
 ## [2.2.2] - 2015-01-13
 ### Added
@@ -65,7 +79,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Client library now uses BitPay's new API
 
-[unreleased]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.2...HEAD
+[unreleased]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.5...HEAD
+[2.2.5]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.4...v2.2.5
+[2.2.4]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.3...v2.2.4
+[2.2.3]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.2...v2.2.3
 [2.2.2]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/bitpay/php-bitpay-client/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/bitpay/php-bitpay-client/compare/v2.1.1...v2.2.0